+++ miaosha-common/miaosha-common-security/src/main/java/tech/tengshe789/miaosha/common/security/config/WebSecurityConfig.java
@@ -1,0 +1,22 @@
+package tech.tengshe789.miaosha.common.security.config;
+
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+/**
+ * @program: -miaosha
+ * @author: tEngSHe789
+ * @create: 2019-03-25 13:00
+ **/
+@EnableWebSecurity
+public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+	@Override
+	protected void configure(HttpSecurity http) throws Exception {
+		http.csrf().disable()
+			.authorizeRequests()
+			.antMatchers("/actuator/**").permitAll()
+			.anyRequest()
+			.authenticated().and().httpBasic();
+	}
+}