+++ miaosha-sys/miaosha-sys-api/src/main/java/tech/tengshe789/miaosha/sys/api/feign/RemoteLogService.java
@@ -8,13 +8,14 @@ import tech.tengshe789.miaosha.common.core.constants.SecurityConstants;
 import tech.tengshe789.miaosha.common.core.constants.ServiceNameConstants;
 import tech.tengshe789.miaosha.common.core.result.Result;
 import tech.tengshe789.miaosha.common.log.entity.SysLog;
+import tech.tengshe789.miaosha.sys.api.feign.factories.RemoteLogServiceFallbackFactory;
 
 /**
  * @program: -miaosha
  * @description: log服务接口
  * @author: <a href="mailto:randyvan007@qq.com">tEngSHe789</a>
  **/
-@FeignClient(value = ServiceNameConstants.UMPS_SERVICE)
+@FeignClient(value = ServiceNameConstants.SYS_SERVICE, fallbackFactory = RemoteLogServiceFallbackFactory.class)
 public interface RemoteLogService {
 	/**
 	 * 保存日志