+++ miaosha-sys/miaosha-sys-api/src/main/java/tech/tengshe789/miaosha/sys/api/feign/fallback/RemoteLogServiceFallbackImpl.java
@@ -1,0 +1,33 @@
+package tech.tengshe789.miaosha.sys.api.feign.fallback;
+
+
+import lombok.Setter;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import tech.tengshe789.miaosha.common.core.result.Result;
+import tech.tengshe789.miaosha.common.log.entity.SysLog;
+import tech.tengshe789.miaosha.sys.api.feign.RemoteLogService;
+
+/**
+ * @author lengleng
+ * @date 2019/2/1
+ */
+@Slf4j
+@Component
+public class RemoteLogServiceFallbackImpl implements RemoteLogService {
+	@Setter
+	private Throwable cause;
+
+	/**
+	 * 保存日志
+	 *
+	 * @param sysLog 日志实体
+	 * @param from   内部调用标志
+	 * @return succes、false
+	 */
+	@Override
+	public Result<Boolean> saveLog(SysLog sysLog, String from) {
+		log.error("feign 插入日志失败", cause);
+		return null;
+	}
+}